refactor: move classes to definition package.
